[snapshot-controller] Use correct duration to wait for CRDs
Ensure the function `ensureCustomResourceDefinitionsExist` returns when
the duration specified in `retryCRDIntervalMax` flag is reached.
This works using by passing a context to the backoff so that we can
timeout with context.WithTimeout.
Also limit the amount of resources returned by the list calls to 0 to
speed up the operations because we only care that the CRD exists.